+++ packages/qwik/src/transformation/index.ts
@@ -1,1 +1,4 @@
 export * from './toCustom$';
+export * from './toLowerCase';
+export * from './toTrimmed';
+export * from './toUpperCase';